Update Installation Guide.md
Added 2.5 and 2.6 to install Python and pip.

Update Installation Guide.md

Made update as per Edit installation doc for git clone step #62

Unit test for delete_organization and if statement

Repair link in documentation

Display a simple notification to the user that they have registered
successfully

Add README detail

Add detail on psycopg2 install

New unit tests for events

Design

New unit tests for job

Pep8 job tests

Refactor job unit tests

Refactor event unit tests + PEP8

Fix print statement for PY3 and PEP8 cleanup

Add coverage config file

Unit tests for shift

Refactor settings for style

Fix spacing in urls.py refactor

Unit tests for volunteer
Unit tests for volunteer

Refactor unit tests for volunteer

Fix import

Fixed import

Datepickers for admin (create job, event, shift)
Datepicker for Create Event

Datepicker for Create Event +

Datepicker for admin create job

Datepicker for create shift

Functionality so that a shift cannot be deleted

Fix volunteer form + PEP8

Location specification for Event

Unique email field

Python 2 to 3

Manage Shift

Timepicker rebased
Timepicker for shift/edit_hours

newlines

Js file

Add location for Shift

Access only for Administrator

Configure access for administrator features

Style No Access page

Volunteer Shift Reminder Emails

Search for the events that start from...to

New line

Fix a bug: volunteer shifts access

Not allowing to create a post dated event

Error message

Add data range for report

Update README

Add timepicker everywhere

Apostrophe for models

Add editorconfig settings

Edit Vagrantfile

Clean up gitignore and document

Enhance UI for creating shifts

Doc string

Administrator report

Shows selected organization

Phone number registration and menu item name

Fix menu for small screens

Show all shifts except signed up

Update Installation Guide.md

I changed the repository link mentioned under the "Clone Project" heading

Update Installation Guide.md

Update Installation Guide.md

Changed "Once the VM booted up" to "Once the VM boots up" in Section "Using Vagrant", line 10.

Update README.md

Rephrased the sentence

Update README.md

Update README.md

Update README.md

Rephrased the sentence

Updated the installation guide

Updated the installation guide

Fixed navigation to default start page when user accesses web address home (i.e. without trailing path to specific page)

Template no fluid

+++ vms/event/services.py
@@ -1,22 +1,58 @@
 from django.core.exceptions import ObjectDoesNotExist
-from event.models import Event
 
-#need to check that this event is not accociated with any jobs, 
-#otherwise the jobs that it is associated with will be cascade deleted
+from event.models import Event
+from job.models import Job
+from shift.models import Shift
+
+
+def event_not_empty(event_id):
+    """ Checks if the event exists and is not empty """
+    result = True
+    event = get_event_by_id(event_id)
+    if not event:
+        result = False
+    return result
+
+
+def get_event_by_shift_id(shift_id):
+    """
+    Returns an event for the shift. Can be helpful for finding an event
+    """
+    is_valid = True
+    result = None
+
+    try:
+        shift = Shift.objects.get(pk=shift_id)
+        event = shift.job.event
+    except ObjectDoesNotExist:
+        is_valid = False
+
+    if is_valid:
+        result = event
+
+    return result
+
+# need to check that this event is not accociated with any jobs,
+# otherwise the jobs that it is associated with will be cascade deleted
 def delete_event(event_id):
 
     result = True
     event = get_event_by_id(event_id)
-    #check if there are currently any jobs associated with this event
-    jobs_in_event = event.job_set.all()
 
-    #can only delete an event if no jobs are currently associated with it
-    if event and (not jobs_in_event):
-        event.delete()
+    if event_not_empty(event_id):
+        # check if there are currently any jobs associated with this event
+        jobs_in_event = event.job_set.all()
+
+        # can only delete an event if no jobs are currently associated with it
+        if event and (not jobs_in_event):
+            event.delete()
+        else:
+            result = False
     else:
         result = False
 
     return result
+
 
 def get_event_by_id(event_id):
 
@@ -30,6 +66,23 @@ def get_event_by_id(event_id):
 
     if is_valid:
         result = event
+
+    return result
+
+
+def get_events_by_date(start_date, end_date):
+    is_valid = True
+    result = None
+    try:
+        event_list = Event.objects.filter(
+                start_date__gte=start_date,
+                start_date__lte=end_date
+                ).order_by('start_date')
+    except ObjectDoesNotExist:
+        is_valid = False
+
+    if is_valid:
+        result = event_list
 
     return result
